r4/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.0.10
Pillow==9.3.0
psycopg2-binary==2.9.5
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
djangorestframework==3.15.0
